requirements.txt
streamlit
pandas
numpy
matplotlib
python-dotenv
groq
openpyxl
